[EM] [feat]: 서보모터 각도 세부 조절

+++ EM/ARM/robot_arm_arduino/main/ServoControl.h
@@ -19,7 +19,7 @@ enum Mode {
 void initializeServos();
 
 // 서보 모터 부드럽게 이동 함수
-void moveServoSmoothly(Servo &servo, int &currentAngle, int targetAngle, const char* axisName);
+void moveServoSmoothly(Servo &servo, float &currentAngle, float targetAngle, const char* axisName);
 
 // 특정 모드로 이동하는 함수
 void setModeAngles(Mode mode);
